Get rid of ConstraintSystemBuilder and TableBuilder. Create builder module
