+++ crates/m3/src/builder/constraint_system.rs
@@ -5,8 +5,6 @@
 // Lookups
 // Columns <- Associated with tables
 
-use std::iter;
-
 // Statement
 // - Channel boundaries
 // - Table sizes
@@ -14,107 +12,59 @@ pub use binius_core::constraint_system::channel::{
 	Boundary, Flush as CompiledFlush, FlushDirection,
 };
 use binius_core::{
-	constraint_system::ConstraintSystem as CompiledConstraintSystem,
+	constraint_system::{channel::ChannelId, ConstraintSystem as CompiledConstraintSystem},
 	oracle::{Constraint, ConstraintPredicate, ConstraintSet, MultilinearOracleSet, OracleId},
 	transparent::step_down::StepDown,
 };
-use binius_field::{underlier::UnderlierType, Field, TowerField};
-use binius_math::{ArithExpr, LinearNormalForm};
+use binius_field::{underlier::UnderlierType, TowerField};
 use binius_utils::checked_arithmetics::log2_ceil_usize;
 use bumpalo::Bump;
 
-use super::error::Error;
-use crate::{types::B128, witness::WitnessIndex};
+use super::{
+	channel::{Channel, Flush},
+	column::{Column, ColumnInfo},
+	error::Error,
+	statement::Instance,
+	table::Table,
+	types::B128,
+	witness::WitnessIndex,
+};
 
-pub type TableId = usize;
-pub type ChannelId = usize;
-pub type ColumnIndex = usize; // REVIEW: Could make these opaque without a constructor, to protect
-							  // access
-
-#[derive(Debug, Clone, Copy)]
-pub struct ColumnShape {
-	pub tower_height: usize,
-	pub pack_factor: usize,
-}
-
-#[derive(Debug)]
-pub struct ColumnId {
-	pub table_id: TableId,
-	pub index: ColumnIndex,
-}
-
-// TODO: Impl Add/Sub/Mul for Col, returning Expr
-
-// feature: TableBuilder needs namespacing
-#[derive(Debug)]
-pub enum Column<F: TowerField = B128> {
-	Committed { tower_level: usize },
-	LinearCombination(LinearNormalForm<F>),
-}
-
-#[derive(Debug)]
-pub struct ColumnInfo<F: TowerField = B128> {
-	pub col: Column<F>,
-	pub name: String,
-	/// This represents how many
-	pub pack_factor: usize,
-	pub is_nonzero: bool,
-}
-
-#[derive(Debug)]
-pub struct ZeroConstraint<F: Field> {
-	pub name: String,
-	pub expr: ArithExpr<F>,
-}
-
-#[derive(Debug)]
-pub struct Table<F: TowerField = B128> {
-	pub id: TableId,
-	pub name: String,
-	pub column_info: Vec<ColumnInfo<F>>,
-	pub flushes: Vec<Flush>,
-	pub zero_constraints: Vec<ZeroConstraint<F>>,
-	/// This indicates whether a table is fixed for constraint system or part of the dynamic trace.
-	///
-	/// Fixed tables are either entirely transparent or committed during a preprocessing step that
-	/// occurs before any statements are proven.
-	pub is_fixed: bool,
-}
-
-#[derive(Debug)]
-pub struct Flush {
-	pub column_indices: Vec<usize>,
-	pub channel_id: ChannelId,
-	pub direction: FlushDirection,
-}
-
-pub struct Instance<F: TowerField = B128> {
-	pub boundaries: Vec<Boundary<F>>,
-	/// Direct index mapping table IDs to the count of rows per table.
-	///
-	/// The table sizes seem like advice values that don't affect the semantic meaning of the
-	/// statement, but we include them in the statement directly. This makes sense because
-	///
-	/// 1. These values affect the control flow of the verification routine.
-	/// 2. These values are necessarily made public.
-	/// 3. For some constraint systems, the verifier does care about the values. For example, the
-	///    statement could be that a VM execution state is reachable within a certain number of
-	///    cycles.
-	pub table_sizes: Vec<usize>,
-}
-
-#[derive(Debug)]
-pub struct Channel {
-	pub name: String,
-}
-
+#[derive(Debug, Default)]
 pub struct ConstraintSystem<F: TowerField = B128> {
 	pub tables: Vec<Table<F>>,
+	pub channels: Vec<Channel>,
 	/// All valid channel IDs are strictly less than this bound.
 	pub channel_id_bound: ChannelId,
 }
 
 impl<F: TowerField> ConstraintSystem<F> {
+	pub fn new() -> Self {
+		Self::default()
+	}
+
+	pub fn add_table(&mut self, name: impl ToString) -> &mut Table<F> {
+		let id = self.tables.len();
+		self.tables.push(Table::new(id, name));
+		self.tables.last_mut().expect("table was just pushed")
+	}
+
+	pub fn add_channel(&mut self, name: impl ToString) -> ChannelId {
+		let id = self.channels.len();
+		self.channels.push(Channel {
+			name: name.to_string(),
+		});
+		id
+	}
+
+	pub fn build_witness<U: UnderlierType>(
+		&self,
+		allocator: &Bump,
+		instance: &Instance,
+	) -> Result<WitnessIndex<U>, Error> {
+		todo!()
+	}
+
 	/// Compiles a [`CompiledConstraintSystem`] for a particular statement.
 	///
 	/// The most important transformation that takes place in this step is creating multilinear
@@ -134,7 +84,7 @@ impl<F: TowerField> ConstraintSystem<F> {
 		let mut table_constraints = Vec::with_capacity(self.tables.len());
 		let mut compiled_flushes = Vec::new();
 		let mut non_zero_oracle_ids = Vec::new();
-		for (table, &count) in iter::zip(&self.tables, &statement.table_sizes) {
+		for (table, &count) in std::iter::zip(&self.tables, &statement.table_sizes) {
 			let Table {
 				id,
 				name,
@@ -214,14 +164,6 @@ impl<F: TowerField> ConstraintSystem<F> {
 			non_zero_oracle_ids,
 			max_channel_id: self.channel_id_bound.saturating_sub(1),
 		})
-	}
-
-	pub fn build_witness<U: UnderlierType>(
-		&self,
-		allocator: &Bump,
-		instance: &Instance,
-	) -> Result<WitnessIndex<U>, Error> {
-		todo!()
 	}
 }
