+++ crates/m3/src/lib.rs
@@ -3,8 +3,4 @@
 pub mod builder;
 pub mod collatz;
 pub mod collatz_high_level;
-pub mod constraint_system;
-pub mod error;
-pub mod types;
 pub mod u32;
-pub mod witness;
