+++ crates/m3/src/u32.rs
@@ -8,11 +8,7 @@ use binius_field::{
 };
 use bytemuck::Pod;
 
-use super::{
-	builder::{Col, TableBuilder},
-	types::B1,
-};
-use crate::witness::TableWitnessIndexSegment;
+use crate::builder::{column::Col, table::Table, types::B1, witness::TableWitnessIndexSegment};
 
 // Concepts:
 //
@@ -53,12 +49,7 @@ pub struct U32AddFlags {
 /// flexibility to adapt method signatures to the specific needs of the component.
 impl U32Add {
 	/// Constructs a new `U32Add` component.
-	pub fn new(
-		table: &mut TableBuilder,
-		xin: Col<B1, 5>,
-		yin: Col<B1, 5>,
-		flags: U32AddFlags,
-	) -> Self {
+	pub fn new(table: &mut Table, xin: Col<B1, 5>, yin: Col<B1, 5>, flags: U32AddFlags) -> Self {
 		let cout = table.add_committed::<B1, 5>("cout");
 		let cout_shl = table.add_shifted("cout_shl", cout, 5, 1, ShiftVariant::LogicalLeft);
